Downloader write errors

+++ libosmscout-client-qt/include/osmscout/AvailableMapsModel.h
@@ -121,7 +121,7 @@ public:
 class OSMSCOUT_CLIENT_QT_API AvailableMapsModelMap : public AvailableMapsModelItem {
   Q_OBJECT
 
-  Q_PROPERTY(qint64 byteSize READ getSize())
+  Q_PROPERTY(quint64 byteSize READ getSize())
   Q_PROPERTY(QString size READ getSizeHuman())
   Q_PROPERTY(QString serverDirectory READ getServerDirectory())
   Q_PROPERTY(QDateTime time READ getCreation())
@@ -129,7 +129,7 @@ class OSMSCOUT_CLIENT_QT_API AvailableMapsModelMap : public AvailableMapsModelIt
 
 private:
   MapProvider provider;
-  size_t size{0};
+  uint64_t size{0};
   QString serverDirectory;
   QDateTime creation;
   int version{-1};
@@ -138,7 +138,7 @@ public:
   inline AvailableMapsModelMap():AvailableMapsModelItem(){};
 
   inline AvailableMapsModelMap(QString name, QList<QString> path, QString description, MapProvider provider,
-                               size_t size, QString serverDirectory, QDateTime creation, int version):
+                               uint64_t size, QString serverDirectory, QDateTime creation, int version):
     AvailableMapsModelItem(name, path, description), provider(provider), size(size), serverDirectory(serverDirectory), 
     creation(creation), version(version) {};
 
@@ -165,7 +165,7 @@ public:
   }
 
   MapProvider getProvider() const;
-  size_t getSize() const;
+  uint64_t getSize() const;
   QString getSizeHuman() const;
   QString getServerDirectory() const;
   QDateTime getCreation() const;
